remove old extensions projects
