--- extension_example2/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "extension_example2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev="d521da1991719760e271457dfe4f9ddf281afbb3" }
-common = { path = "../common" }
-extensions2 = { path = "../extensions2" }
-kit = { path = "../kit" }
-once_cell = "1.17.1"
