--- extension_example2/src/style.css
@@ -1,6 +1,0 @@
-body,
-html,
-#main,
-#app-wrap {
-  background: orangered;
-}
