--- extensions2/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "extensions2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dioxus = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da1991719760e271457dfe4f9ddf281afbb3" }
-libloading = "0.7.4"
-warp = { git = "https://github.com/Satellite-im/Warp", rev = "6e50802ba6e4d5c1ad10969d9a3a8aa176d47432" }
-derive_more = "0.99"
-ipc-channel = "0.16.0"
-
-[build-dependencies]
-rustc_version = "0.4.0"
