+++ native_extensions/emoji_selector/Cargo.toml
@@ -15,7 +15,7 @@ dioxus-desktop = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da19
 ] }
 dioxus-hooks = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da1991719760e271457dfe4f9ddf281afbb3" }
 dioxus-html = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da1991719760e271457dfe4f9ddf281afbb3" }
-extensions2 = { path = "../../extensions2" }
+extensions = { path = "../../extensions" }
 common = { path = "../../common" }
 kit = { path = "../../kit" }
 emojis = "0.5.2"
