+++ ui/Cargo.toml
@@ -17,7 +17,7 @@ dioxus-core = { git = "https://github.com/DioxusLabs/dioxus", rev = "d521da19917
 fermi = { git = "https://github.com/DioxusLabs/dioxus", rev = "22e71a71bdcdc03c3ae83ae1c3b3fb5417ebaa80" }
 kit = { path = "../kit" }
 common = { path = "../common" }
-extensions2 = { path = "../extensions2" }
+extensions = { path = "../extensions" }
 tokio-util = { version = "0.7", features = ["full"] }
 arboard = "3.1.0"
 humansize = "2.1.3"
