+++ ui/src/extension_browser/extensions.scss
