--- ui/src/extensions/mod.rs
@@ -1,1 +1,0 @@
-pub mod browser;
